链路信息查询匹配接口名,上传接口改为上传文件.

+++ traffic-service/api/src/main/java/com/oc/hawk/traffic/entrypoint/api/dto/TraceItemDTO.java
@@ -11,7 +11,7 @@ public class TraceItemDTO {
     private String path;
     private String method;
     private String dstWorkload;
-    private Integer latency;
+    private Long latency;
     private String spanId;
     private String responseCode;
     private String startTime;
